Product Backlog Item 115:Cargar imagenes en el Crear Beneficio

+++ Src/PointEx.Web/Areas/Admin/Controllers/SectionItemController.cs
@@ -82,7 +82,7 @@ namespace PointEx.Web.Areas.Admin.Controllers
 
             int pageTotal = 0;
             var benefits = _benefitService.GetAll("CreatedDate", "DESC", filters.CategoryId, filters.TownId,
-                filters.ShopId, filters.Criteria, (int)BenefitStatusEnum.Approved, filters.Page, DefaultPageSize, out pageTotal);
+                filters.ShopId, filters.Criteria, BenefitStatusEnum.Approved, filters.Page, DefaultPageSize, out pageTotal);
 
             var pagedList = new StaticPagedList<BenefitDto>(benefits, filters.Page, DefaultPageSize, pageTotal);
