+++ Src/PointEx.Web/Controllers/BenefitController.cs
@@ -31,7 +31,7 @@ namespace PointEx.Web.Controllers
         {
             int pageTotal;
 
-            var benefits = _benefitService.GetAll("CreatedDate", "DESC", filters.CategoryId, filters.TownId, filters.ShopId, filters.Criteria, (int)BenefitStatusEnum.Approved, filters.Page, DefaultPageSize, out pageTotal);
+            var benefits = _benefitService.GetAll("CreatedDate", "DESC", filters.CategoryId, filters.TownId, filters.ShopId, filters.Criteria, BenefitStatusEnum.Approved, filters.Page, DefaultPageSize, out pageTotal);
 
             var pagedList = new StaticPagedList<BenefitDto>(benefits, filters.Page, DefaultPageSize, pageTotal);
 
@@ -63,8 +63,6 @@ namespace PointEx.Web.Controllers
             {
                 throw new HttpException(404, "Not found");
             }
-
-
         }
     }
 }
